added flight form operations (add, edit, delete) added a script to import dataset for airlines and airports to support autosuggestions

+++ server/src/services/flight-service.ts
@@ -1,16 +1,29 @@
 import { prisma } from '../prisma/client.js';
 import type { FlightFormData } from '../controllers/flight-controller.js';
+import type { CreateFlightInput, UpdateFlightInput } from '../schemas/flight-schema.js';
 
-const create = async (data: FlightFormData) => {
+
+const create = async (tripId: string, data: CreateFlightInput) => {
+
+    const isTripExists = await prisma.trip.findUnique({
+        where: { id: tripId }
+    });
+
+    if (!isTripExists) {
+        throw new Error("Trip not found");
+    }
+
     return prisma.flight.create({
         data: {
-            tripId: data.tripId,
+            tripId,
             flightNumber: data.flightNumber,
             airline: data.airline,
             departure: data.departure,
             arrival: data.arrival,
             from: data.from,
-            to: data.to
+            to: data.to,
+            departureTimezoneId: data.departureTimezoneId || 'UTC',
+            arrivalTimezoneId: data.arrivalTimezoneId || 'UTC'
         }
     });
 };
@@ -24,7 +37,16 @@ const getByTripId = async (tripId: string) => {
     });
 };
 
-const update = async (flightId: string, data: Partial<FlightFormData>) => {
+const update = async (flightId: string, data: UpdateFlightInput) => {
+    
+    const isFlightExists = await prisma.flight.findUnique({
+        where: { id: flightId }
+    });
+
+    if (!isFlightExists) {
+        throw new Error("Flight not found");
+    }
+
     return prisma.flight.update({
         where: { id: flightId },
         data: {
@@ -33,12 +55,31 @@ const update = async (flightId: string, data: Partial<FlightFormData>) => {
             ...(data.departure !== undefined && { departure: data.departure }),
             ...(data.arrival !== undefined && { arrival: data.arrival }),
             ...(data.from !== undefined && { from: data.from }),
-            ...(data.to !== undefined && { to: data.to })
+            ...(data.to !== undefined && { to: data.to }),
+            ...(data.departureTimezoneId !== undefined && { departureTimezoneId: data.departureTimezoneId }),
+            ...(data.arrivalTimezoneId !== undefined && { arrivalTimezoneId: data.arrivalTimezoneId })
         }
     });
 };
 
-const deleteFlight = async (flightId: string) => {
+const deleteFlight = async (flightId: string, tripId: string) => {
+
+    const isTripExists = await prisma.trip.findUnique({
+        where: { id: tripId }
+    });
+
+    if (!isTripExists) {
+        throw new Error("Trip not found");
+    }
+
+    const isFlightExists = await prisma.flight.findUnique({
+        where: { id: flightId }
+    });
+
+    if (!isFlightExists) {
+        throw new Error("Flight not found");
+    }
+
     return prisma.flight.delete({
         where: { id: flightId }
     });
